requirements.txt
flask==2.2.3
werkzeug==2.2.3
flask-cors==3.0.10
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.2.2
psycopg2-binary==2.9.1
python-dotenv==0.19.1
itsdangerous==2.1.2
jinja2==3.1.2
click==8.1.3
markupsafe==2.1.2
SQLAlchemy==1.4.46